requirements.txt
backports.functools-lru-cache==1.4
cycler==0.10.0
docopt==0.6.2
matplotlib==2.1.0
numpy==1.13.3
pandas==0.21.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
six==1.11.0
subprocess32==3.2.7
